Building process is now 100% handled by rollup
