+++ config/tsconfig.json
@@ -4,20 +4,16 @@
         "module": "esnext",
         "moduleResolution": "node",
         "lib": ["esNext"],
+        "rootDir": "../src",
         "outDir": "../dist/esnext",
+
         "forceConsistentCasingInFileNames": true,
         "strict": true,
         "declaration": true,
         "declarationMap": true,
         "declarationDir": "../dist/types",
         "resolveJsonModule": true, // To import json
-        "plugins": [
-          {
-            "name": "typescript-tslint-plugin",
-            "alwaysShowRuleFailuresAsWarnings": false
-          }
-        ]
     },
     "include": ["../src"], // files to include during transpilation
-    "compileOnSave": false
 }
+
