+++ package.json
@@ -19,9 +19,9 @@
   "main": "./dist/bundles/meilisearch.cjs.js",
   "module": "./dist/bundles/meilisearch.esm.js",
   "browser": "./dist/bundles/meilisearch.umd.js",
-  "typings": "./dist/types/types.d.ts",
-  "types": "./dist/types/types.d.ts",
-  "jsnext:main": "./dist/esnext/meilisearch.js",
+  "typings": "./src/types.d.ts",
+  "types": "./src/types.d.ts",
+  "jsnext:main": "./dist/bundles/meilisearch.esm.js",
   "sideEffects": false,
   "repository": {
     "type": "git",
@@ -29,8 +29,8 @@
   },
   "scripts": {
     "cleanup": "shx rm -rf dist/",
-    "build": "yarn cleanup && tsc -p config && rollup -c config/rollup.config.js && rollup -c config/rollup.config.js --environment NODE_ENV:production",
-    "postbuild": "yarn size && yarn typingsheader",
+    "build": "yarn cleanup && rollup -c config/rollup.config.js && rollup -c config/rollup.config.js --environment NODE_ENV:production",
+    "postbuild": "yarn size",
     "test": "jest -c ./config/jest.config.js --runInBand",
     "test:watch": "yarn test --watch",
     "test:coverage": "yarn test --coverage",
@@ -83,6 +83,7 @@
     "pretty-bytes": "^5.3.0",
     "rollup": "^2.2.0",
     "rollup-plugin-terser": "^5.3.0",
+    "rollup-plugin-typescript2": "^0.27.0",
     "shx": "^0.3.2",
     "ts-jest": "^25.3.0",
     "tslint": "^6.1.0",
